combined the two appointment-related cards into a single unified card
