+++ lib/features/home/service/components/appointment_card.dart
@@ -1,4 +1,4 @@
-/// Number of appointments card widget.
+/// Combined appointment card widget.
 //
 // Time-stamp: <Friday 2025-02-21 08:30:05 +1100 Graham Williams>
 //
@@ -27,28 +27,41 @@ library;
 
 import 'package:flutter/material.dart';
 
+import 'package:audioplayers/audioplayers.dart';
 import 'package:intl/intl.dart';
 import 'package:markdown_tooltip/markdown_tooltip.dart';
 
 import 'package:healthpod/theme/card_style.dart';
 
-/// A widget that shows the number of upcoming medical appointments.
-///
-/// This card displays a summary of the number of appointments scheduled
-/// for the future, with an edit button to add or manage appointments.
+// Add this global variable if it doesn't exist elsewhere
+bool transportAudioIn = false;
 
-class NumberAppointments extends StatefulWidget {
-  const NumberAppointments({super.key});
+/// A widget that displays both next appointment details and appointment summary.
+///
+/// This component combines the functionality of showing the next appointment
+/// details and the total number of upcoming appointments in a single card.
+
+class AppointmentCard extends StatefulWidget {
+  const AppointmentCard({super.key});
 
   @override
-  State<NumberAppointments> createState() => _NumberAppointmentsState();
+  State<AppointmentCard> createState() => _AppointmentCardState();
 }
 
-class _NumberAppointmentsState extends State<NumberAppointments> {
-  // Appointment data.
+class _AppointmentCardState extends State<AppointmentCard> {
+  /// Status of playing of the audio.
+  bool _isPlaying = false;
+  final AudioPlayer _audioPlayer = AudioPlayer();
 
-  String title = 'Numbers for Medical Appointments';
-  String summary = 'Only one appointment in the future';
+  // Appointment data
+  String title = 'Medical Appointments';
+  String subtitle = 'Next Appointment Details';
+  DateTime appointmentDate = DateTime(2023, 3, 13, 14, 30);
+  String location = 'Gurriny Yealamucka';
+  bool needsTransport = true;
+  String transportPhone = '(07) 4226 4100';
+  String transportNote = '(only during office hours)';
+  bool useClinicBus = true;
   List<Map<String, dynamic>> appointments = [
     {
       'title': 'General Checkup',
@@ -58,8 +71,49 @@ class _NumberAppointmentsState extends State<NumberAppointments> {
     }
   ];
 
-  /// Opens a dialog to manage appointments.
+  /// Toggles the audio playback state.
+  Future<void> _toggleAudio() async {
+    if (_isPlaying) {
+      await _audioPlayer.stop();
+      setState(() {
+        _isPlaying = false;
+        transportAudioIn = false;
+      });
+    } else {
+      if (!transportAudioIn) {
+        await _audioPlayer.play(AssetSource('audio/transport_eligibility.mp3'));
+        setState(() {
+          _isPlaying = !_isPlaying;
+          transportAudioIn = true;
+        });
+      }
+    }
+  }
 
+  /// Handles the completion of audio playback.
+  void _onAudioComplete() {
+    setState(() {
+      _isPlaying = false;
+      transportAudioIn = false;
+    });
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    _audioPlayer.onPlayerComplete.listen((event) {
+      _onAudioComplete();
+    });
+  }
+
+  @override
+  void dispose() {
+    _audioPlayer.dispose();
+    transportAudioIn = false;
+    super.dispose();
+  }
+
+  /// Opens a dialog to manage all appointments.
   void _manageAppointments() {
     showDialog(
       context: context,
@@ -111,8 +165,6 @@ class _NumberAppointmentsState extends State<NumberAppointments> {
                       alignment: Alignment.center,
                       child: ElevatedButton.icon(
                         onPressed: () {
-                          // Show dialog to add new appointment.
-
                           _showAddAppointmentDialog(context, setState);
                         },
                         icon: const Icon(Icons.add),
@@ -120,15 +172,11 @@ class _NumberAppointmentsState extends State<NumberAppointments> {
                       ),
                     ),
                     const SizedBox(height: 16),
-                    // Import/Export buttons.
-
                     Row(
                       mainAxisAlignment: MainAxisAlignment.spaceBetween,
                       children: [
                         OutlinedButton.icon(
                           onPressed: () {
-                            // TODO: Implement import functionality.
-
                             Navigator.pop(context);
                             ScaffoldMessenger.of(context).showSnackBar(
                               const SnackBar(
@@ -141,8 +189,6 @@ class _NumberAppointmentsState extends State<NumberAppointments> {
                         ),
                         OutlinedButton.icon(
                           onPressed: () {
-                            // TODO: Implement export functionality.
-
                             Navigator.pop(context);
                             ScaffoldMessenger.of(context).showSnackBar(
                               const SnackBar(
@@ -167,16 +213,12 @@ class _NumberAppointmentsState extends State<NumberAppointments> {
                 ),
                 ElevatedButton(
                   onPressed: () {
-                    // Update the state in the card widget.
-
                     this.setState(() {
-                      if (appointments.isEmpty) {
-                        summary = 'No upcoming appointments';
-                      } else if (appointments.length == 1) {
-                        summary = 'Only one appointment in the future';
-                      } else {
-                        summary =
-                            '${appointments.length} appointments scheduled';
+                      // Update the next appointment if there are any appointments
+                      if (appointments.isNotEmpty) {
+                        final nextAppointment = appointments.first;
+                        appointmentDate = nextAppointment['date'];
+                        location = nextAppointment['location'];
                       }
                     });
                     Navigator.pop(context);
@@ -192,7 +234,6 @@ class _NumberAppointmentsState extends State<NumberAppointments> {
   }
 
   /// Shows a dialog to add a new appointment.
-
   void _showAddAppointmentDialog(
       BuildContext context, void Function(void Function()) parentSetState) {
     final titleController = TextEditingController();
@@ -305,7 +346,6 @@ class _NumberAppointmentsState extends State<NumberAppointments> {
                   return;
                 }
 
-                // Add the new appointment to the list
                 parentSetState(() {
                   appointments.add({
                     'title': titleController.text,
@@ -328,7 +368,10 @@ class _NumberAppointmentsState extends State<NumberAppointments> {
   @override
   Widget build(BuildContext context) {
     return Container(
-      width: 400,
+      constraints: const BoxConstraints(
+        maxWidth: 400,
+        minHeight: 300,
+      ),
       padding: const EdgeInsets.all(16.0),
       decoration: getHomeCardDecoration(context),
       child: Column(
@@ -357,11 +400,128 @@ class _NumberAppointmentsState extends State<NumberAppointments> {
           ),
           const SizedBox(height: 8),
           Text(
-            summary,
+            appointments.isEmpty
+                ? 'No upcoming appointments'
+                : appointments.length == 1
+                    ? 'Only one appointment in the future'
+                    : '${appointments.length} appointments scheduled',
             style: const TextStyle(fontSize: 14),
           ),
+          const SizedBox(height: 16),
+          if (appointments.isNotEmpty) ...[
+            Text(
+              subtitle,
+              style: const TextStyle(
+                fontSize: 16,
+                fontWeight: FontWeight.w500,
+              ),
+            ),
+            const SizedBox(height: 16),
+            // Date
+            _buildInfoRow(
+              'Date:',
+              'Monday, ${DateFormat('d MMMM').format(appointmentDate)}',
+            ),
+            const SizedBox(height: 8),
+            // Time
+            _buildInfoRow(
+              'Time:',
+              DateFormat('h:mm a').format(appointmentDate),
+            ),
+            const SizedBox(height: 8),
+            // Location
+            _buildInfoRow('Where:', location),
+            const SizedBox(height: 16),
+            // Transport
+            if (useClinicBus)
+              Row(
+                children: [
+                  const Icon(Icons.directions_bus, color: Colors.green),
+                  const SizedBox(width: 8),
+                  const Text(
+                    'Clinic Bus:',
+                    style: TextStyle(fontWeight: FontWeight.bold),
+                  ),
+                  const Icon(Icons.check, color: Colors.green),
+                ],
+              ),
+            if (needsTransport) ...[
+              const SizedBox(height: 16),
+              const Text(
+                'Need help with transport?',
+                style: TextStyle(fontWeight: FontWeight.bold),
+              ),
+              const SizedBox(height: 8),
+              Row(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  const Icon(Icons.phone, size: 20),
+                  const SizedBox(width: 8),
+                  Expanded(
+                    child: Text.rich(
+                      TextSpan(
+                        children: [
+                          TextSpan(
+                            text: 'Call $transportPhone ',
+                            style: const TextStyle(fontWeight: FontWeight.bold),
+                          ),
+                          TextSpan(
+                            text: transportNote,
+                            style: const TextStyle(fontStyle: FontStyle.italic),
+                          ),
+                          const TextSpan(
+                            text: ' to change or request transport.',
+                          ),
+                        ],
+                      ),
+                    ),
+                  ),
+                  SizedBox(
+                    width: 32,
+                    height: 32,
+                    child: MarkdownTooltip(
+                      message: _isPlaying
+                          ? '**Stop** audio'
+                          : '**Play** audio explanation',
+                      child: IconButton(
+                        icon: Icon(
+                          _isPlaying ? Icons.stop : Icons.volume_up,
+                          color: _isPlaying ? Colors.red : Colors.blue,
+                          size: 20,
+                        ),
+                        onPressed: _toggleAudio,
+                        padding: EdgeInsets.zero,
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ],
+          ],
         ],
       ),
+    );
+  }
+
+  /// Helper method to build consistent information rows.
+  Widget _buildInfoRow(String label, String value) {
+    return Row(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        SizedBox(
+          width: 60,
+          child: Text(
+            label,
+            style: const TextStyle(
+              fontWeight: FontWeight.bold,
+            ),
+          ),
+        ),
+        Expanded(
+          child: Text(value),
+        ),
+      ],
     );
   }
 }
